friend_finder/requirements.txt
torch
torchvision
facenet-pytorch
gradio
pillow
numpy
